added loginCard component
